The beam momentum spread for the e^+/e^- beam of ELPH GeV-gamma was introduced. Reference: 	GeV-γ 解析ノート HD No. 439E 電子・陽電子ビームライン IV 	T. Ishikawa 	23rd October 2020, Revised 03rd November 2020 	in Japanese
To enable the beam momentum spread, you need to write
  /INTT/beam/beamSmearing true
in a Geant4 macro. The default value is false.
You can select
  - beammline: -30 or -23(defalt), -30 degree line provides higher momentum beam with larger momentum spread with respect to the -23 degree beamline
    For example, write
      /INTT/beam/beamLine -23
      /INTT/beam/beamLine -30
    in a Geant4 macro.
  - production target: Au_20um(default), W_200um, and Cu_8mm are available. Thicker targets generate higher intensity beams with larger momentum spread, especially at low momentum due to multiple-scattering by the target.
    For example, you can write
    	/INTT/beam/productionTarget Au_20um
	/INTT/beam/productionTarget W_200um
	/INTT/beam/productionTarget Cu_8mm
    om a Geant4 macro.
  - position restriction of beam: The momentum spread is strongly correlated to horizontal beam position because the bending angle of the beam by the magnet creates the horizontal beam position distribution. If one selects beams in certain horizontal area, you can get a much smaller momentum spread. The possible choiscs are 4. Read the reference for more details.
    For example, you can specify the restriction by
    	/INTT/beam/positionRestriction 0
	/INTT/beam/positionRestriction 1
	/INTT/beam/positionRestriction 2
	/INTT/beam/positionRestriction 3
    in a Geant4 macro. 0(default) means no restriction, 1, 2, 3 mean |x|<4.5cm (weak restriction), |x|<3.0cm (middle), |x|<1.5cm (strong), respectively.
    It's OK to use 0 for the moment.

A new class ELPHEBeam was made. EDRunAction class keeps an instance of the class as a static object. The beam momentum and the momentum spread are obtained in EDPrimaryGeneratorAction by accessing the static instance.
To use ELPHEBeam class, you need to use ROOT. If ROOT is already available in your environment, just doing cmake again should be all you have to do. Ask me for help (genki.nukazuka@riken.jp).

Data was taken from the reference. Only the case of Al target was done. Once data is extracted from the reference and input to TSV (tab-separated-values) file, you can generate a ROOT file. Then name of the file should be Al_20um.root, W_200um.root, or Cu_8mm.root. The rest of the targets should be done by Japanese colleagues.